337-subscriptionField: Readd react select styles
